requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djoser==2.2.3
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.2.0
python-dotenv==1.1.1
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.32.4